% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{createDotPlot}
\alias{createDotPlot}
\title{Create Dot Plot}
\usage{
createDotPlot(currentData)
}
\value{
a todo
}
\description{
todo
}
\examples{

## Create data frame containing required information to generate graph
dataInfo <- data.frame(gene_name=c(rep("Myc", 4), rep("Kras", 4)), 
    Mean=c(0.349, 0.232, 0.052, 0.0634, 0.1607, 0.5099, 0.1860, 0.1325), 
    Percentage=c(45.85, 45.02, 7.07, 7.64, 31.87, 70.76, 17.17, 19.74),
    ClusterName=rep(c("Fibroblasts", "EMT-like Cells", 
                      "T and NK Cells", "Myeloid Cells"), 2),
    stringAsFactor=FALSE)
    
## Create Dot Plot
dplot <- createDotPlot(currentData=dataInfo)   
print(dplot)
                  
}
\author{
Astrid Deschênes
}
\keyword{internal}
