% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\encoding{UTF-8}
\name{createUI}
\alias{createUI}
\title{Define the user interface (UI)}
\usage{
createUI()
}
\value{
a user interface (UI) definition that can be passed to 
the \code{shinyUI} function.
}
\description{
todo
}
\examples{

## todo

}
\author{
Astrid Deschênes
}
\keyword{internal}
