% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\encoding{UTF-8}
\name{createUIDotplotPanel}
\alias{createUIDotplotPanel}
\title{Define the user interface (UI) panel for the dot plot section}
\usage{
createUIDotplotPanel()
}
\value{
a user interface (UI) definition that can be passed to 
the \code{shinyUI} function.
}
\description{
todo
}
\examples{

## todo

}
\author{
Astrid Deschênes
}
\keyword{internal}
