% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseScPDACshiny.R
\docType{package}
\encoding{UTF-8}
\name{mouseScPDACshiny-package}
\alias{mouseScPDACshiny-package}
\alias{mouseScPDACshiny}
\title{mouseScPDACshiny: Interactive Generator of Dot Plot using Elyada 
et al 2019 mouse KPC Single-Cell Dataset}
\source{
Elyada E. et al. Cross-Species Single-Cell Analysis of Pancreatic Ductal 
Adenocarcinoma Reveals Antigen-Presenting Cancer-Associated Fibroblasts. 
Cancer Discov. 2019 Aug;9(8):1102-1123. doi: 10.1158/2159-8290.CD-19-0094. 
Epub 2019 Jun 13. PMID: 31197017; PMCID: PMC6727976.
}
\description{
This package provides a graphical user interface (GUI) to generate 
a dot plot showing the mean expression and ratio of cell expressing  
genes selected by user. The data is from Elyada et al. 2019 mouse KPC 
single-cell dataset.
}
\seealso{
\describe{
\item{\code{\link{startApp}}}{ for launching an interactive session}
}
}
\author{
Astrid Deschênes

Maintainer:
Astrid Deschênes <adeschen@hotmail.com>
}
