% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouseScPDACshiny.R
\docType{data}
\name{seAll}
\alias{seAll}
\title{This dataset contains the information from the Elyada et all 2019 mouse 
single-cell KPC dataset. The information is stored in a 
\code{SummarizedExperiment} with one \code{assays} containing the mean 
expression per cluster for each gene and a second  \code{assays} containing 
the percentage of cells with expression per cluster for each gene.}
\format{
The \code{SummarizedExperiment} with one \code{assays} containing 
the mean expression per cluster for each gene and a second  \code{assays} 
containing the percentage of cells with expression per cluster for each 
gene.
}
\usage{
data(seAll)
}
\value{
The \code{SummarizedExperiment} with one \code{assays} containing 
the mean expression per cluster for each gene and a second  \code{assays} 
containing the percentage of cells with expression per cluster for each 
gene.
}
\description{
This dataset contains the information from the Elyada et all 2019 mouse 
single-cell KPC dataset. The information is stored in a 
\code{SummarizedExperiment} with one \code{assays} containing the mean 
expression per cluster for each gene and a second  \code{assays} containing 
the percentage of cells with expression per cluster for each gene.
}
\examples{

## Load dataset
data(seAll)

if (required(SummarizedExperiment)) {
    ## Load required library
    library(SummarizedExperiment)
    
    ## Mean expression of the 3 first genes (rows) in each cluster (columns)
    assays(seAll)$mean[1:3, ]
}


}
\seealso{
\describe{
\item{\code{\link{startApp}}}{ for launching an interactive session}
}
}
\keyword{datasets}
