% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\encoding{UTF-8}
\name{server}
\alias{server}
\title{Server function}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{todo}

\item{output}{todo}

\item{session}{todo

@return todo}
}
\description{
TOTO
}
\keyword{internal}
