% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\encoding{UTF-8}
\name{startApp}
\alias{startApp}
\title{Start the shiny application}
\usage{
startApp(...)
}
\arguments{
\item{...}{todo}
}
\value{
An object that represents the shiny application. Printing the 
object or passing it to \code{runApp()} will run the application.
}
\description{
This function starts the mouseKPCscElyada2019shiny shiny 
application.
}
\examples{

## Required library 
library(shiny)

## Only run this example in interactive R sessions
if (interactive()) {
    app <- startApp()
    runApp(app)
}

## Another way to start the application
## Only run this example in interactive R sessions
if (interactive()) {
    startApp()
}

}
\author{
Astrid Deschênes
}
