% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\encoding{UTF-8}
\name{ui_download}
\alias{ui_download}
\title{Create the download section of the user interface (UI)}
\usage{
ui_download()
}
\value{
a section of an user interface (UI) definition that can be passed to 
the \code{shinyUI} function.
}
\description{
todo
}
\examples{

## todo

}
\author{
Astrid Deschênes
}
\keyword{internal}
